# axum/openapi.yaml
openapi: 3.1.0
info:
  title: Ogonek API
  description: ''
  contact:
    name: API Support
    email: danila.volkov@noxlovette.com
  license:
    name: ''
  version: 1.6.0
servers:
- url: https://api.ogonek.app
  description: Production server
- url: http://localhost:3000
  description: Local development server
paths:
...
  /api/v1/s3/abort:
    post:
      tags:
      - Task
      summary: Cancel multipart upload
      operationId: abort_multipart_upload
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbortMultipartRequest'
        required: true
      responses:
        '200':
          description: Upload aborted successfully
        '400':
          description: Bad request
        '404':
          description: File not found
      security:
      - api_key: []
  /api/v1/s3/complete:
    post:
      tags:
      - Task
      summary: Complete a part of the upload
      operationId: complete_multipart_upload
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteMultipartRequest'
        required: true
      responses:
        '201':
          description: Upload completed successfully
        '400':
          description: Bad request
        '404':
          description: File not found
      security:
      - api_key: []
  /api/v1/s3/init:
    post:
      tags:
      - Task
      summary: Multipart upload init endpoint
      operationId: init_multipart_upload
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitUploadRequest'
        required: true
      responses:
        '200':
          description: Multipart upload initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultipartUploadInit'
        '400':
          description: Bad request
        '404':
          description: Parent folder not found
      security:
      - api_key: []
  /api/v1/s3/presigned/{encoded_key}:
    get:
      tags:
      - Task
      summary: Presigns url for frontend download
      operationId: fetch_presigned_url
      parameters:
      - name: encoded_key
        in: path
        description: File ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Presigned URL generated successfully
        '401':
          description: Unauthorized
      security:
      - api_key: []
  /api/v1/s3/{id}:
    delete:
      tags:
      - Task
      summary: Deletes file
      operationId: delete_file
      parameters:
      - name: id
        in: path
        description: File ID
        required: true
        schema:
          type: string
      responses:
        '204':
          description: File deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: File not found
      security:
      - api_key: []
...
components:
  schemas:
    AbortMultipartRequest:
      type: object
      required:
      - uploadId
      - fileId
      - s3Key
      properties:
        fileId:
          type: string
        s3Key:
          type: string
        uploadId:
          type: string
...
    CompleteMultipartRequest:
      type: object
      required:
      - uploadId
      - fileId
      - s3Key
      - parts
      properties:
        fileId:
          type: string
        parts:
          type: array
          items:
            $ref: '#/components/schemas/CompletedPart'
        s3Key:
          type: string
        uploadId:
          type: string
    CompletedPart:
      type: object
      required:
      - partNumber
      - etag
      properties:
        etag:
          type: string
        partNumber:
          type: integer
          format: int32
...
    InitUploadRequest:
      type: object
      required:
      - fileName
      - contentType
      - fileSize
      - totalParts
      properties:
        contentType:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
          format: int64
        parentId:
          type:
          - string
          - 'null'
        taskId:
          type:
          - string
          - 'null'
        totalParts:
          type: integer
          format: int32
...
    MultipartUploadInit:
      type: object
      required:
      - uploadId
      - fileId
      - s3Key
      - parts
      properties:
        fileId:
          type: string
        parts:
          type: array
          items:
            $ref: '#/components/schemas/PartUploadUrl'
        s3Key:
          type: string
        uploadId:
          type: string
...
    PartUploadUrl:
      type: object
      required:
      - partNumber
      - url
      properties:
        partNumber:
          type: integer
          format: int32
        url:
          type: string
...
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: X-API-Key
...